requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==4.2.5
django-crispy-forms==2.0
Pillow==10.0.1
sqlparse==0.4.4
